+++ common/26.1.2/src/main/java/com/ksoichiro/mcmod/fancyicecream/item/GlowBerryIceCream.java
@@ -1,0 +1,13 @@
+package com.ksoichiro.mcmod.fancyicecream.item;
+
+import net.minecraft.world.item.Item;
+
+public class GlowBerryIceCream extends Item {
+    public GlowBerryIceCream(String itemName) {
+        super(new ItemProperties(itemName)
+                .food(new FoodBuilder()
+                        .nutrition(4)
+                        .saturationMod(0.3F)
+                        .build()));
+    }
+}