+++ common/26.2/src/main/java/com/ksoichiro/mcmod/fancyicecream/entity/decoration/IceCreamCup.java
@@ -1,0 +1,34 @@
+package com.ksoichiro.mcmod.fancyicecream.entity.decoration;
+
+import com.ksoichiro.mcmod.fancyicecream.FancyIceCream;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.Identifier;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.decoration.HangingEntity;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+
+public class IceCreamCup extends IceCreamStand {
+    public IceCreamCup(EntityType<? extends HangingEntity> entityType, Level level) {
+        super(entityType, level);
+    }
+
+    public IceCreamCup(Level level, BlockPos pos, Direction direction, Direction placedDirection) {
+        super(getCupEntityType(), level, pos, direction, placedDirection);
+    }
+
+    private static EntityType<? extends HangingEntity> getCupEntityType() {
+        return (EntityType<? extends HangingEntity>) BuiltInRegistries.ENTITY_TYPE
+                .getValue(Identifier.fromNamespaceAndPath(FancyIceCream.MOD_ID, "ice_cream_cup"));
+    }
+
+    @Override
+    public ItemStack getFrameItemStack() {
+        Item item = BuiltInRegistries.ITEM.getValue(
+                Identifier.fromNamespaceAndPath(FancyIceCream.MOD_ID, "ice_cream_cup"));
+        return new ItemStack(item);
+    }
+}